+++ src/Http/Controllers/Settings/Users/BaseUserSettingsController.php
@@ -5,18 +5,20 @@ declare(strict_types=1);
 namespace CraftCms\Cms\Http\Controllers\Settings\Users;
 
 use CraftCms\Cms\Cp\Data\NavItem;
-use Inertia\Inertia;
 
 use function CraftCms\Cms\cp_url;
 use function CraftCms\Cms\t;
 
 abstract class BaseUserSettingsController
 {
-    public function __construct()
+    /**
+     * @return NavItem[]
+     */
+    protected function subnav(): array
     {
         $path = request()->craftPath();
 
-        Inertia::share('subnav', [
+        return [
             new NavItem()
                 ->label(t('User Groups'))
                 ->url(cp_url('settings/users'))
@@ -29,7 +31,7 @@ abstract class BaseUserSettingsController
                 ->label(t('Settings'))
                 ->url(cp_url('settings/users/settings'))
                 ->selected($path === 'settings/users/settings'),
-        ]);
+        ];
     }
 
     protected function crumbs(string $title, ?string $url = null): array