+++ workbench/app/Providers/TypeScriptTransformerServiceProvider.php
@@ -12,6 +12,7 @@ use CraftCms\Cms\Gql\Data\GqlToken;
 use CraftCms\Cms\Http\ViewModels\FieldEditViewModel;
 use CraftCms\Cms\Http\ViewModels\UserPermissionsViewModel;
 use CraftCms\Cms\Http\ViewModels\UserPreferencesViewModel;
+use CraftCms\Cms\Http\ViewModels\UserSignInProvidersViewModel;
 use CraftCms\Cms\Image\Data\ImageTransform;
 use CraftCms\Cms\Route\Data\Route;
 use CraftCms\Cms\Update\Data\Updates;
@@ -53,6 +54,7 @@ class TypeScriptTransformerServiceProvider extends TypeScriptTransformerApplicat
                     UserPermissionsViewModel::class,
                     UserPreferencesViewModel::class,
                     UserSettings::class,
+                    UserSignInProvidersViewModel::class,
                 ],
                 [
                     new EnumTransformer,